# tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatcherStatusResponse": {
        "properties": {
          "dirty_count": {
            "title": "Dirty Count",
            "type": "integer"
          },
          "last_event_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event At"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "requires_reconciliation": {
            "title": "Requires Reconciliation",
            "type": "boolean"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "state",
          "dirty_count",
          "requires_reconciliation",
          "last_event_at",
          "updated_at"
        ],
        "title": "WatcherStatusResponse",
        "type": "object"
      },
      "WatchersListResponse": {
        "properties": {
          "watchers": {
            "items": {
              "$ref": "#/components/schemas/WatcherStatusResponse"
            },
            "title": "Watchers",
            "type": "array"
          }
        },
        "required": [
          "watchers"
        ],
        "title": "WatchersListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminSessionCookie": {
        "in": "cookie",
        "name": "workflows_admin_session",
        "type": "apiKey"
      },
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "CsrfToken": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "workflows-mcp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/v1/watchers": {
      "get": {
        "operationId": "list_watchers_api_admin_v1_watchers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchersListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "List Watchers",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/watchers/{project_id}": {
      "get": {
        "operationId": "get_watcher_status_api_admin_v1_watchers__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatcherStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Get Watcher Status",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/watchers/{project_id}/disable": {
      "post": {
        "operationId": "disable_watcher_api_admin_v1_watchers__project_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatcherStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Disable Watcher",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/watchers/{project_id}/pause": {
      "post": {
        "operationId": "pause_watcher_api_admin_v1_watchers__project_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatcherStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Pause Watcher",
        "tags": [
          "admin-v1"
        ]
      }
    },
    "/api/admin/v1/watchers/{project_id}/resume": {
      "post": {
        "operationId": "resume_watcher_api_admin_v1_watchers__project_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatcherStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AdminSessionCookie": [],
            "CsrfToken": []
          }
        ],
        "summary": "Resume Watcher",
        "tags": [
          "admin-v1"
        ]
      }
    },
...
    "/api/events/v1/watchers": {
      "get": {
        "operationId": "watcher_events_api_events_v1_watchers_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "summary": "Watcher Events",
        "tags": [
          "events-v1"
        ]
      }
    },
...
  }
}